+++ surmod/gpytorch_gaussian_process.py
@@ -407,11 +407,11 @@ class GPSurrogate:
         mean, std = self.predict(self.x_test, include_nugget=include_nugget)
         y_true = self.y_test.squeeze(-1).cpu().numpy()
 
-        mse = float(np.mean((y_true - mean) ** 2))
-        rmse = float(np.sqrt(mse))
+        mse = np.mean((y_true - mean) ** 2)
+        rmse = np.sqrt(mse)
         lower = mean - 1.96 * std
         upper = mean + 1.96 * std
-        coverage = float(np.mean((y_true >= lower) & (y_true <= upper)))
+        coverage = np.mean((y_true >= lower) & (y_true <= upper))
 
         return {
             "mse": mse,
